wado-compiler: extract import_adapter submodule from cm_binding
Move synthesize_adapter (the WASI import binding generator), the
flat-Result discriminant lifter, the make_binding_function helper
shared with the export side, and the binding_func_name builder into
`synthesis/cm_binding/import_adapter.rs`.
